# Conflicts:
#	src/main/pl/edu/agh/to/core/CoreSystem.scala
#	src/main/pl/edu/agh/to/core/IslandUtils.scala
#	src/main/pl/edu/agh/to/core/TestApplication.scala
#	src/main/pl/edu/agh/to/util/OptionalUtils.scala

+++ src/main/pl/edu/agh/to/agent/Agent.java
@@ -25,7 +25,7 @@ public class Agent {
 
     }
 
-    public double getFitness() {
+    public int getFitness() {
         return operator.evaluation(this);
     }
